Revert 134060 (broke compilation on ChromeOS builder) - gdata: GetAvailableSpace should return int64
We should use 64-bit integer for the arguments bytes_total and bytes_used in
the GetAvailableSpace callback.

Updating unit tests with >4GB sizes to verify that the sizes are not clamped.

BUG=chromium-os:30011
TEST=Verify that gdata file system unit tests pass.


Review URL: http://codereview.chromium.org/10174033

TBR=hshi@google.com
Review URL: https://chromiumcodereview.appspot.com/10198021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@134063 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: fa8083d5da85c7f7cae0bffb5f6eeb14e0bbca22

+++ chrome/browser/chromeos/gdata/gdata_file_system.h
@@ -79,8 +79,8 @@ typedef base::Callback<void(base::PlatformFileError error,
 
 // Used to get available space for the account from GData.
 typedef base::Callback<void(base::PlatformFileError error,
-                            int64 bytes_total,
-                            int64 bytes_used)>
+                            int bytes_total,
+                            int bytes_used)>
     GetAvailableSpaceCallback;
 
 // Callback type for DocumentServiceInterface::ResumeUpload.
